chore(client/ts): port widgets/buttons/launcher

+++ src/public/app/widgets/buttons/launcher/abstract_launcher.ts
@@ -1,14 +1,20 @@
 import shortcutService from "../../../services/shortcuts.js";
 import attributesService from "../../../services/attributes.js";
 import OnClickButtonWidget from "../onclick_button.js";
+import type FNote from "../../../entities/fnote.js";
+import type FAttribute from "../../../entities/fattribute.js";
+import type { EventData } from "../../../components/app_context.js";
+import type { AttributeRow } from "../../../services/load_results.js";
 
-export default class AbstractLauncher extends OnClickButtonWidget {
-    constructor(launcherNote) {
+export default abstract class AbstractLauncher extends OnClickButtonWidget {
+
+    protected launcherNote: FNote;
+
+    constructor(launcherNote: FNote) {
         super();
 
         this.class("launcher-button");
 
-        /** @type {FNote} */
         this.launcherNote = launcherNote;
 
         for (const label of launcherNote.getOwnedLabels("keyboardShortcut")) {
@@ -16,22 +22,20 @@ export default class AbstractLauncher extends OnClickButtonWidget {
         }
     }
 
-    launch() {
-        throw new Error("Abstract implementation");
-    }
+    abstract launch(): void;
 
-    bindNoteShortcutHandler(labelOrRow) {
+    bindNoteShortcutHandler(labelOrRow: FAttribute | AttributeRow) {
         const namespace = labelOrRow.attributeId;
 
-        if (labelOrRow.isDeleted) {
+        if ("isDeleted" in labelOrRow && labelOrRow.isDeleted) {
             // only applicable if row
             shortcutService.removeGlobalShortcut(namespace);
-        } else {
+        } else if (labelOrRow.value) {
             shortcutService.bindGlobalShortcut(labelOrRow.value, () => this.launch(), namespace);
         }
     }
 
-    entitiesReloadedEvent({ loadResults }) {
+    entitiesReloadedEvent({ loadResults }: EventData<"entitiesReloaded">) {
         for (const attr of loadResults.getAttributeRows()) {
             if (attr.noteId === this.launcherNote.noteId && attr.type === "label" && attr.name === "keyboardShortcut") {
                 this.bindNoteShortcutHandler(attr);
